Create ChopSticks.py
I made the game chopsticks in python. It is not the best code but it works. I will be making single player with a really crappy AI, It will just be a random number gen seeing if it will attack or split, left hand or right and stuff like that. Have fun playing it. P.S if you print in into repl.it and share it with a friend you can both play on the same console with different computers as with the collab effect it is a shared console.